requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.11.0
trl>=0.8.0
bitsandbytes>=0.43.0

# Quantization / Merging
scipy

# Evaluation
nltk>=3.8
sqlparse>=0.5.0
sqlglot>=23.0

# Experiment Tracking
wandb>=0.16.0

# Inference / Demo
gradio>=4.20.0

# Data
pandas>=2.0.0
pyarrow

# Config
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
rich>=13.0.0
tqdm
